method added to scrape links to news topics in nav

+++ news_content_snatch.py
@@ -1,12 +1,14 @@
 import bs4 as bs
 import urllib2 as request
+import os
+
+URL_HEAD = 'http://www.bbc.com'
 
 
 def get_full_content(link):
     content_local = 'news/'
-    head = 'http://www.bbc.com'
     try:
-        sauce = request.urlopen(head + link).read()
+        sauce = request.urlopen(URL_HEAD + link).read()
     except request.HTTPError:
         print '404 encountered'
         return
@@ -18,8 +20,52 @@ def get_full_content(link):
     for paragraph in article.find_all(['p', 'h2']):
         paragraphs.append(paragraph.get_text())
     news_path = content_local + link.split('/')[2]
-    print news_path
-    with open(news_path, 'w') as writer:
-        for p in paragraphs:
-            print p
-            writer.write(p.encode('utf8'))
+
+    if not os.path.isfile(news_path):
+        with open(news_path, 'w') as writer:
+            for p in paragraphs:
+                writer.write(p.encode('utf8'))
+        print 'news entry added'
+
+
+def get_nav_links():
+    sauce = request.urlopen(URL_HEAD + '/news').read()
+    soup = bs.BeautifulSoup(sauce, 'lxml')
+    body = soup.body
+    nav = body.find('div', class_='navigation--wide')
+
+    nav_links = {}
+
+    for topic in nav.find_all('li'):
+        if topic.get('class') == ' invisible':
+            break
+        topic_link = topic.find('a').get('href')
+
+        if topic_link == '/news/video_and_audio/headlines':
+            continue
+        if topic_link == '/news/magazine':
+            break
+
+        sauce = request.urlopen(URL_HEAD + topic_link).read()
+        soup = bs.BeautifulSoup(sauce, 'lxml')
+        body = soup.body
+
+        try:
+            sub_nav = body.find('nav', class_='navigation-wide-list--secondary').find('ul')
+        except AttributeError:
+            nav_links[topic_link] = []
+            continue
+        if sub_nav is None:
+            nav_links[topic_link] = []
+            continue
+
+        sub_links = []
+        for sub_topic in sub_nav.find_all('li', class_=None):
+            sub_topic_link = sub_topic.find('a').get('href')
+            if sub_topic_link == 'http://www.bbc.co.uk/news/business/market_data'\
+                    or sub_topic_link == '/news/business/markets':
+                continue
+            sub_links.append(sub_topic_link)
+
+        nav_links[topic_link] = sub_links
+    return nav_links
